+++ example-app/actions-outputs/server-actions-for-next-pages/context.js
@@ -1,0 +1,2 @@
+import { wrapGetServerSideProps as _wrapGetServerSideProps } from "server-actions-for-next-pages/dist/context-internal";
+module.exports = require('./dist/context');